requirements.txt
django
djangorestframework
django-cors-headers
gunicorn
djangorestframework-jwt
django-filter
Pillow
django-rest-swagger
django-post_office
